requirements.txt
fastapi
uvicorn
psycopg2-binary
supabase
python-multipart
python-magic
python-dotenv
apscheduler
pytest
httpx
pydantic-settings
slowapi